Rename MessageSender to MultiprocessingSender; move to its own module
